+++ orury-client/src/main/java/org/orury/client/gym/interfaces/message/GymMessage.java
@@ -1,4 +1,4 @@
-package org.orury.client.gym.converter.message;
+package org.orury.client.gym.interfaces.message;
 
 import lombok.Getter;
