+++ orury-client/src/main/java/org/orury/client/gym/interfaces/response/GymResponse.java
@@ -1,7 +1,7 @@
-package org.orury.client.gym.converter.response;
+package org.orury.client.gym.interfaces.response;
 
 import org.orury.domain.global.constants.Constants;
-import org.orury.domain.gym.dto.GymDto;
+import org.orury.domain.gym.domain.dto.GymDto;
 
 import java.time.DayOfWeek;
 import java.util.List;
