+++ orury-client/src/main/java/org/orury/client/gym/interfaces/response/GymsResponse.java
@@ -1,6 +1,6 @@
-package org.orury.client.gym.converter.response;
+package org.orury.client.gym.interfaces.response;
 
-import org.orury.domain.gym.dto.GymDto;
+import org.orury.domain.gym.domain.dto.GymDto;
 
 public record GymsResponse(
         Long id,
