+++ orury-common/build.gradle
@@ -4,9 +4,6 @@ dependencies {
     implementation 'org.flywaydb:flyway-core'
     implementation 'org.flywaydb:flyway-mysql'
 
-    //S3
-    implementation 'com.amazonaws:aws-java-sdk-s3:1.12.268'
-
     //Jasypt
     implementation group: 'com.github.ulisesbocchio', name: 'jasypt-spring-boot-starter', version: '3.0.4'
