+++ orury-domain/src/main/java/org/orury/domain/gym/domain/entity/GymLike.java
@@ -1,4 +1,4 @@
-package org.orury.domain.gym.db.model;
+package org.orury.domain.gym.domain.entity;
 
 import jakarta.persistence.EmbeddedId;
 import jakarta.persistence.Entity;
