+++ orury-domain/src/main/java/org/orury/domain/gym/domain/entity/GymLikePK.java
@@ -1,4 +1,4 @@
-package org.orury.domain.gym.db.model;
+package org.orury.domain.gym.domain.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Embeddable;
